+++ examples/get_model.py
@@ -12,39 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-"""Install the given Rel source in the given database"""
+"""Fetch the given Rel model from the given database"""
 
 from argparse import ArgumentParser
-import json
-from os import path
-from urllib.request import HTTPError
-from railib import api, config, show
+from railib import api, config
+from show_error import show_error
 
 
-# Reeturns the file basename without extension.
-def _sansext(fname: str) -> str:
-    return path.splitext(path.basename(fname))[0]
-
-
-def run(database: str, engine: str, fname: str, profile: str):
-    sources = {}
-    with open(fname) as fp:
-        sources[_sansext(fname)] = fp.read()  # source name => source
+@show_error
+def run(database: str, engine: str, model: str, profile: str):
     cfg = config.read(profile=profile)
     ctx = api.Context(**cfg)
-    rsp = api.install_source(ctx, database, engine, sources)
-    print(json.dumps(rsp, indent=2))
+    rsp = api.get_model(ctx, database, engine, model)
+    print(rsp)
 
 
 if __name__ == "__main__":
     p = ArgumentParser()
     p.add_argument("database", type=str, help="database name")
     p.add_argument("engine", type=str, help="engine name")
-    p.add_argument("file", type=str, help="source file")
+    p.add_argument("model", type=str, help="model name")
     p.add_argument("-p", "--profile", type=str,
                    help="profile name", default="default")
     args = p.parse_args()
-    try:
-        run(args.database, args.engine, args.file, args.profile)
-    except HTTPError as e:
-        show.http_error(e)
+    run(args.database, args.engine, args.model, args.profile)