+++ examples/list_models.py
@@ -12,18 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-"""Delete the given Rel source from the given database."""
+"""List the models installed in the given database"""
 
 from argparse import ArgumentParser
 import json
-from urllib.request import HTTPError
-from railib import api, config, show
+from railib import api, config
+from show_error import show_error
 
 
-def run(database: str, engine: str, source: str, profile: str):
+@show_error
+def run(database: str, engine: str, profile: str):
     cfg = config.read(profile=profile)
     ctx = api.Context(**cfg)
-    rsp = api.delete_source(ctx, database, engine, source)
+    rsp = api.list_models(ctx, database, engine)
     print(json.dumps(rsp, indent=2))
 
 
@@ -31,11 +32,6 @@ if __name__ == "__main__":
     p = ArgumentParser()
     p.add_argument("database", type=str, help="database name")
     p.add_argument("engine", type=str, help="engine name")
-    p.add_argument("source", type=str, help="source name")
-    p.add_argument("-p", "--profile", type=str,
-                   help="profile name", default="default")
+    p.add_argument("-p", "--profile", type=str, help="profile name", default="default")
     args = p.parse_args()
-    try:
-        run(args.database, args.engine, args.source, args.profile)
-    except HTTPError as e:
-        show.http_error(e)
+    run(args.database, args.engine, args.profile)