+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## v0.6.3
 
+* Rename `source` to `model`
 * Rename api.list_edb to api.list_edbs
-* Support create/list/get/delete oauth clients
-* Support create/disable users
-* Support delete databases by name
+* Implement create/list/get/delete oauth clients
+* Implement create/disable users
+* Implement delete database by name
 
 ## v0.6.2